requirements.txt
streamlit
torch
torchvision
transformers
huggingface-hub
Pillow
numpy
opencv-python
matplotlib
scipy
scikit-image
groq
python-dotenv
torchxrayvision
monai
einops
timm
albumentations
